Add files via upload
all the countries across the globe have adopted to means of digital payments. And with the increased volume of digital payments, hacking has become pretty common event wherein the hacker can try to hack your details just with your phone number linked to your bank account. However, there is data with some anonymized variables based on which one can predict whether the hack is going to happen.
your work is to build a predictive model that can identify a pattern in these variables and suggest that a hack is going to happen so that the cybersecurity can somehow stop it before it actually happens. you have to predict the column "MULTIPLE_OFFENSE".
*******************************************************************************************************************************
COLUMN	                        DESCRIPTION
INCIDENT_ID	                UNIQUE IDENTIFIER FOR AN INCIDENT LOG
DATE	                        DATE OF INCIDENT OCCURRENCE
X_1 - X_15	                ANONYMIZED LOGGING PARAMETERS
MULTIPLE_OFFENSE	[TARGET], INDICATES WHETHER THE INCIDENT WAS A HACK

****************************************************************************************************************************************